PIM: avoid compiler warnings in 32 bit
gcc warns about a "long int" vs. "unsigned int" mismatch when
compiling for x86. All of that is in debugging output printfs.
Cast the index_t into long int to avoid that.

+++ src/dbus/server/pim/filtered-view.cpp
@@ -151,7 +151,7 @@ void FilteredView::addIndividual(int parentIndex, const IndividualData &data)
             if (newEndIndex > index &&
                 !m_filter->isIncluded(newEndIndex)) {
                 const IndividualData *data = m_parent->getContact(m_local2parent.back());
-                SE_LOG_DEBUG(NULL, NULL, "%s: removed at #%ld/%ld to make room for new entry", getName(), newEndIndex - 1, m_local2parent.size());
+                SE_LOG_DEBUG(NULL, NULL, "%s: removed at #%ld/%ld to make room for new entry", getName(), (long)(newEndIndex - 1), (long)m_local2parent.size());
                 m_local2parent.pop_back();
                 m_removedSignal(newEndIndex - 1, *data);
                 // Iterator might have pointed to removed entry, which may have
@@ -160,10 +160,10 @@ void FilteredView::addIndividual(int parentIndex, const IndividualData &data)
             }
 
             m_local2parent.insert(it, parentIndex);
-            SE_LOG_DEBUG(NULL, NULL, "%s: added at #%ld/%ld", getName(), index, m_local2parent.size());
+            SE_LOG_DEBUG(NULL, NULL, "%s: added at #%ld/%ld", getName(), (long)index, (long)m_local2parent.size());
             m_addedSignal(index, data);
         } else {
-            SE_LOG_DEBUG(NULL, NULL, "%s: not added at #%ld/%ld because outside of result range", getName(), index, m_local2parent.size());
+            SE_LOG_DEBUG(NULL, NULL, "%s: not added at #%ld/%ld because outside of result range", getName(), (long)index, (long)m_local2parent.size());
         }
     }
 }
@@ -189,7 +189,7 @@ void FilteredView::removeIndividual(int parentIndex, const IndividualData &data)
 
     if (found) {
         size_t index = it - m_local2parent.begin();
-        SE_LOG_DEBUG(NULL, NULL, "%s: removed at #%ld/%ld", getName(), index, m_local2parent.size());
+        SE_LOG_DEBUG(NULL, NULL, "%s: removed at #%ld/%ld", getName(), (long)index, (long)m_local2parent.size());
         m_local2parent.erase(it);
         m_removedSignal(index, data);
         // Try adding more contacts from the parent if there is room now.
@@ -209,11 +209,11 @@ void FilteredView::modifyIndividual(int parentIndex, const IndividualData &data)
         size_t index = it - m_local2parent.begin();
         if (matches) {
             // Still matched, merely pass on modification signal.
-            SE_LOG_DEBUG(NULL, NULL, "%s: modified at #%ld/%ld", getName(), index, m_local2parent.size());
+            SE_LOG_DEBUG(NULL, NULL, "%s: modified at #%ld/%ld", getName(), (long)index, (long)m_local2parent.size());
             m_modifiedSignal(index, data);
         } else {
             // Removed.
-            SE_LOG_DEBUG(NULL, NULL, "%s: removed at #%ld/%ld due to modification", getName(), index, m_local2parent.size());
+            SE_LOG_DEBUG(NULL, NULL, "%s: removed at #%ld/%ld due to modification", getName(), (long)index, (long)m_local2parent.size());
             int candidate = m_local2parent.back() + 1;
             m_local2parent.erase(it);
             m_removedSignal(index, data);
@@ -224,7 +224,7 @@ void FilteredView::modifyIndividual(int parentIndex, const IndividualData &data)
         size_t index = it - m_local2parent.begin();
         if (m_filter->isIncluded(index)) {
             m_local2parent.insert(it, parentIndex);
-            SE_LOG_DEBUG(NULL, NULL, "%s: added at #%ld/%ld due to modification", getName(), index, m_local2parent.size());
+            SE_LOG_DEBUG(NULL, NULL, "%s: added at #%ld/%ld due to modification", getName(), (long)index, (long)m_local2parent.size());
             m_addedSignal(index, data);
         }
     } else {
